+++ CHANGELOG.md
@@ -18,6 +18,10 @@
 - `offsetTop` in Height Viewport component only applies if its element's top position is smaller than half the viewport
 - Larger width for form width x-small
 
+### Fixed
+
+- Offcanvas will keep scroll position if anchor link was clicked
+
 ## 3.0.0 beta 25 (June 9, 2017)
 
 ### Added